server/requirements.txt
# Core Packages
Django==3.0.7
djangorestframework==3.11.0
django-cors-headers==3.4.0
djangorestframework-simplejwt==4.4.0
django-fsm==2.7.0
drf-yasg==1.17.1
channels==2.4.0
channels-redis==3.0.1
python-decouple==3.3

# Coding standards
pylint==2.6.0

# Server
gunicorn==20.0.4
daphne==2.5.0

# Testing
coverage==5.2.1

# Debuggers
bpython==0.19
django-debug-toolbar==2.2
django-extensions==3.0.0
